Create README.md
In this week's independent project, we are going to wear Pauline's shoes, who works as an Operations Manager at XYZ company. As an essential component of the onboarding process, she is tasked with creating employee IDs as well as the official employee email.

Here's the process she has to do to create the employee ID and email,

 Log in to https://excelcult.com/hr-on-boarding-process-form/ (Links to an external site.) website. This website allows her to enter employees' details in order to generate the employee ID and email.
Enters the newly employed employee's details. The details include First Name, Last Name, Date of birth (DOB), Designation, Mobile number, and gender.
Accept the terms and conditions
Submit the form details
Get the employee ID and email generated after submitting the form.
Your challenge is to help Pauline automate this process so that she can concentrate on more important tasks in her department. Your goal is to save the generated employee ID and Email in a NotePad application and save the file name as Employee details. 

Additionally, you will be required to ask the user to enter their details then use those details to fill out the form on the website.